+++ src/test/java/org/folio/sidecar/integration/keycloak/filter/KeycloakJwtFilterTest.java
@@ -256,6 +256,55 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
   }
 
   @Test
+  void filter_positive_publicMissingToken() {
+    var requestHeaders = headers(Collections.emptyMap());
+    var routingContext = routingContext(scRoutingEntry("not-system"), rc -> {
+      when(rc.request()).thenReturn(request);
+      when(request.headers()).thenReturn(requestHeaders);
+    });
+
+    var result = keycloakJwtFilter.applyFilter(routingContext);
+
+    assertThat(result.succeeded()).isTrue();
+    assertThat(result.result()).isEqualTo(routingContext);
+  }
+
+  @Test
+  void filter_negative_wildcardMissingToken() {
+    var requestHeaders = headers(Collections.emptyMap());
+    var routingContext = routingContext(scRoutingEntry("not-system", "*"), rc -> {
+      when(rc.request()).thenReturn(request);
+      when(request.headers()).thenReturn(requestHeaders);
+    });
+
+    var result = keycloakJwtFilter.applyFilter(routingContext);
+
+    assertThat(result.succeeded()).isFalse();
+    assertThat(result.cause())
+      .isInstanceOf(UnauthorizedException.class)
+      .hasMessage("Failed to find JWT in request");
+  }
+
+  @Test
+  void filter_negative_wildcardInvalidToken() {
+    var requestHeaders = headers(Map.of(TOKEN, AUTH_TOKEN));
+    var routingContext = routingContext(scRoutingEntry("not-system", "*"), rc -> {
+      when(rc.request()).thenReturn(request);
+      when(request.headers()).thenReturn(requestHeaders);
+    });
+
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.failedFuture(
+      new UnauthorizedException("Failed to parse JWT", new ParseException("Failed to parse JWT, invalid offset"))));
+
+    var result = keycloakJwtFilter.applyFilter(routingContext);
+
+    assertThat(result.succeeded()).isFalse();
+    assertThat(result.cause())
+      .isInstanceOf(UnauthorizedException.class)
+      .hasMessage("Failed to parse JWT");
+  }
+
+  @Test
   void filter_negative_tokenMismatch() {
     var token1 = "dG9rZW4tb25l";
     var token2 = "dG9rZW4tdHdv";